+++ src/main/java/de/derivo/neo2rdf/conversion/Neo4jToRDFConverter.java
@@ -70,9 +70,7 @@ public abstract class Neo4jToRDFConverter {
         this.relationshipTypeIDToStr = indexedSchema.getRelationshipTypeIDToStr();
 
         Neo4jToRDFMapperBuilder builder = new Neo4jToRDFMapperBuilder(config.getBasePrefix(), this.indexedSchema);
-        builder
-                .setReifyRelationships(config.isReifyRelationships())
-                .setReificationVocabulary(config.getReificationVocabulary());
+        builder.setReificationVocabulary(config.getReificationVocabulary());
         this.neo4jToRDFMapper = builder.build();
 
         nodeProcessor = new NodeToRDFConverter(neoStores, this, config);