+++ src/main/java/de/derivo/neo2rdf/conversion/config/ConversionConfigBuilder.java
@@ -4,6 +4,8 @@ import de.derivo.neo2rdf.util.ReificationVocabulary;
 import de.derivo.neo2rdf.util.SequenceConversionType;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
 
 public class ConversionConfigBuilder {
     private String basePrefix = "https://www.example.org#";
@@ -16,6 +18,7 @@ public class ConversionConfigBuilder {
     private boolean derivePropertyHierarchyByRelationshipSubsetCheck = false;
     private boolean reifyRelationships = true;
     private boolean reifyOnlyRelationshipsWithProperties = false;
+    private List<String> relationshipTypeReificationBlacklist = new ArrayList<>();
     private File schemaOutputPath = null;
 
     public ConversionConfigBuilder() {
@@ -88,6 +91,15 @@ public class ConversionConfigBuilder {
         return this;
     }
 
+    public List<String> getRelationshipTypeReificationBlacklist() {
+        return relationshipTypeReificationBlacklist;
+    }
+
+    public ConversionConfigBuilder setRelationshipTypeReificationBlacklist(List<String> relationshipTypeReificationBlacklist) {
+        this.relationshipTypeReificationBlacklist = relationshipTypeReificationBlacklist;
+        return this;
+    }
+
     public ConversionConfig build() {
         ConversionConfig config = new ConversionConfig();
         config.reificationVocabulary = reificationVocabulary;
@@ -101,6 +113,9 @@ public class ConversionConfigBuilder {
         config.derivePropertyHierarchyByRelationshipSubsetCheck = derivePropertyHierarchyByRelationshipSubsetCheck;
         config.schemaOutputPath = schemaOutputPath;
         config.reifyRelationships = reifyRelationships;
+        config.relationshipTypeReificationBlacklist = relationshipTypeReificationBlacklist;
         return config;
     }
+
+
 }