+++ src/test/java/de/derivo/neo2rdf/Neo4jToRDFReificationTests.java
@@ -14,6 +14,8 @@ import java.util.List;
 
 public class Neo4jToRDFReificationTests {
 
+    private static final String basePrefix = "https://www.example.org#";
+
     @RegisterExtension
     public static final RDFStoreTestExtension storeTestExtension = new RDFStoreTestExtension(TestUtil.getResource(
             "neo4j-multi-relationships"));
@@ -21,13 +23,15 @@ public class Neo4jToRDFReificationTests {
 
     @Test
     void testRDFReification() {
-        ConversionConfig config = new ConversionConfigBuilder().setReificationVocabulary(ReificationVocabulary.RDF_REIFICATION).build();
+        ConversionConfig config = new ConversionConfigBuilder()
+                .setBasePrefix(basePrefix)
+                .setReificationVocabulary(ReificationVocabulary.RDF_REIFICATION).build();
         storeTestExtension.convertAndImportIntoStore("multi-relationship-reified-with-rdf.ttl", config);
         try (TupleQueryResult bindingSets = storeTestExtension.executeQuery("""
                 PREFIX rdf:        <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
-                PREFIX rdfs:       <http://www.w3.org/2000/01/rdf-schema#>
-                SELECT ?statement ?s ?p ?o WHERE {
-                    ?statement rdf:subject ?s ; rdf:predicate ?p ; rdf:object ?o .
+                PREFIX :        <https://www.example.org#>
+                SELECT ?statement WHERE {
+                    ?statement rdf:subject ?s ; rdf:predicate :WATCHED ; rdf:object ?o .
                 }
                 """)) {
             Assertions.assertEquals(12, bindingSets.stream().count());
@@ -36,12 +40,15 @@ public class Neo4jToRDFReificationTests {
 
     @Test
     void testOWLReification() {
-        ConversionConfig config = new ConversionConfigBuilder().setReificationVocabulary(ReificationVocabulary.OWL_REIFICATION).build();
+        ConversionConfig config = new ConversionConfigBuilder()
+                .setBasePrefix(basePrefix)
+                .setReificationVocabulary(ReificationVocabulary.OWL_REIFICATION).build();
         storeTestExtension.convertAndImportIntoStore("multi-relationship-reified-with-owl.ttl", config);
         try (TupleQueryResult bindingSets = storeTestExtension.executeQuery("""
                 PREFIX owl:        <http://www.w3.org/2002/07/owl#>
-                SELECT ?statement ?s ?p ?o WHERE {
-                    ?statement owl:annotatedSource ?s ; owl:annotatedProperty ?p ; owl:annotatedTarget ?o .
+                PREFIX :        <https://www.example.org#>
+                SELECT ?statement WHERE {
+                    ?statement owl:annotatedSource ?s ; owl:annotatedProperty :WATCHED ; owl:annotatedTarget ?o .
                 }
                 """)) {
             Assertions.assertEquals(12, bindingSets.stream().count());
@@ -51,13 +58,15 @@ public class Neo4jToRDFReificationTests {
     @Test
     void testReificationOnlyOfRelationshipsWithNeo4jProperties() {
         ConversionConfig config = new ConversionConfigBuilder().setReificationVocabulary(ReificationVocabulary.OWL_REIFICATION)
+                .setBasePrefix(basePrefix)
                 .setReifyOnlyRelationshipsWithProperties(true)
                 .build();
         storeTestExtension.convertAndImportIntoStore("reify-only-relationships-with-properties.ttl", config);
         try (TupleQueryResult bindingSets = storeTestExtension.executeQuery("""
                 PREFIX owl:        <http://www.w3.org/2002/07/owl#>
-                SELECT ?statement ?s ?p ?o WHERE {
-                    ?statement owl:annotatedSource ?s ; owl:annotatedProperty ?p ; owl:annotatedTarget ?o .
+                PREFIX :        <https://www.example.org#>
+                SELECT ?statement WHERE {
+                    ?statement owl:annotatedSource ?s ; owl:annotatedProperty :WATCHED ; owl:annotatedTarget ?o .
                 }
                 """)) {
             Assertions.assertEquals(6, bindingSets.stream().count());
@@ -67,15 +76,15 @@ public class Neo4jToRDFReificationTests {
     @Test
     void testDiscardReification() {
         ConversionConfig config = new ConversionConfigBuilder().setReificationVocabulary(ReificationVocabulary.OWL_REIFICATION)
-                .setBasePrefix("https://www.example.org#")
+                .setBasePrefix(basePrefix)
                 .setReifyRelationships(false)
                 .build();
         storeTestExtension.convertAndImportIntoStore("do-not-reify-relationships.ttl", config);
         try (TupleQueryResult bindingSets = storeTestExtension.executeQuery("""
                 PREFIX owl:        <http://www.w3.org/2002/07/owl#>
                 PREFIX :        <https://www.example.org#>
-                SELECT ?statement ?s ?p ?o WHERE {
-                    ?statement owl:annotatedSource ?s ; owl:annotatedProperty ?p ; owl:annotatedTarget ?o .
+                SELECT ?statement WHERE {
+                    ?statement owl:annotatedSource ?s ; owl:annotatedProperty :WATCHED ; owl:annotatedTarget ?o .
                 }
                 """)) {
             Assertions.assertEquals(0, bindingSets.stream().count());
@@ -90,6 +99,38 @@ public class Neo4jToRDFReificationTests {
                 """)) {
             List<BindingSet> results = bindingSets.stream().toList();
             Assertions.assertEquals(7, results.size());
+        }
+    }
+
+    @Test
+    void testReificationBlacklist() {
+        ConversionConfig config = new ConversionConfigBuilder().setReificationVocabulary(ReificationVocabulary.OWL_REIFICATION)
+                .setBasePrefix(basePrefix)
+                .setRelationshipTypeReificationBlacklist(List.of("WATCHED", "NON-EXISTENT-1", "NON-EXISTENT-2", "READ"))
+                .build();
+        storeTestExtension.convertAndImportIntoStore("do-not-reify-relationships.ttl", config);
+        try (TupleQueryResult bindingSets = storeTestExtension.executeQuery("""
+                PREFIX owl:        <http://www.w3.org/2002/07/owl#>
+                PREFIX :           <https://www.example.org#>
+                SELECT ?statement WHERE {
+                    VALUES ?p {
+                        :WATCHED :READ
+                    }
+                    ?statement owl:annotatedSource ?s ; owl:annotatedProperty ?p ; owl:annotatedTarget ?o .
+                }
+                """)) {
+            Assertions.assertEquals(0, bindingSets.stream().count());
+        }
+
+        try (TupleQueryResult bindingSets = storeTestExtension.executeQuery("""
+                PREFIX owl:        <http://www.w3.org/2002/07/owl#>
+                PREFIX :           <https://www.example.org#>
+                SELECT ?s ?p ?o WHERE {
+                    ?s :KNOWS ?o .
+                }
+                """)) {
+            List<BindingSet> results = bindingSets.stream().toList();
+            Assertions.assertEquals(4, results.size());
         }
     }
 }